pyproject.toml
[project]
name = "influencer-discovery-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "chromium>=0.0.0",
    "langchain-google-genai>=4.2.0",
    "lxml>=6.0.2",
    "pandas>=3.0.0",
    "playwright>=1.57.0",
    "playwright-stealth>=2.0.1",
    "python-dotenv>=1.2.1",
]
